fastapi-clean-api/requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
pytest
httpx
email-validator

passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
passlib[bcrypt]